Chain TIM1, TIM2 and TIM3 to create a 48-bit system clock

+++ src/sdvx_animation/mod.rs
@@ -6,5 +6,5 @@ pub use sdvx_fade_animation::*;
 
 pub trait SdvxAnimation {
     fn new() -> Self;
-    fn tick(&mut self, status: &SdvxStatus, current_tick: u32) -> [u8; BCM_LED_COUNT];
+    fn tick(&mut self, status: &SdvxStatus, current_tick: u64) -> [u8; BCM_LED_COUNT];
 }
